third_party/blink/web_tests/http/tests/devtools/elements/styles-4/disable-last-property-without-semicolon-expected.txt

Verifies that formatter adds a semicolon when enabling property.


Running: initFormattedStyle

Running: testFormattedDisableLast
raw cssText:
{
          color: red;
          /* margin: 0; */
      }

Running: testFormattedInsertEnd
raw cssText:
{
          color: red;
          /* margin: 0; */
          endProperty: endValue;
      }

Running: testFormattedEnable
raw cssText:
{
          color: red;
          margin: 0;
          endProperty: endValue;
      }
